pyproject.toml
[tool.poetry]
name = "lavava2025-bot"
version = "0.1.0"
description = ""
authors = ["codenilson <fcodenilson@gmail.com>"]
license = "MIT License"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.5.2"
python-dotenv = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
